Refactor: moved files and resolve import problems
