+++ api/src/main/java/com/egomogo/api/global/adapter/webclient/dto/CoordinateDto.java
@@ -1,4 +1,4 @@
-package com.egomogo.api.service.dto.restaurant;
+package com.egomogo.api.global.adapter.webclient.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
