--- domain/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'domain'
