--- domain/src/main/java/com/egomogo/domain/DomainApplication.java
@@ -1,13 +1,0 @@
-package com.egomogo.domain;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DomainApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DomainApplication.class, args);
-    }
-
-}
