+++ domain/src/main/java/com/egomogo/domain/entity/Restaurant.java
@@ -1,8 +1,8 @@
-package com.egomogo.api.service.entity;
+package com.egomogo.domain.entity;
 
 import com.egomogo.api.global.util.Generator;
-import com.egomogo.api.service.entity.base.BaseAuditEntity;
-import com.egomogo.api.service.type.CategoryType;
+import com.egomogo.domain.entity.base.BaseAuditEntity;
+import com.egomogo.domain.type.CategoryType;
 import jakarta.persistence.*;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
