+++ domain/src/main/java/com/egomogo/domain/repository/MenuRepository.java
@@ -1,10 +1,11 @@
-package com.egomogo.api.service.repository;
+package com.egomogo.domain.repository;
 
-import com.egomogo.api.service.entity.Menu;
+import com.egomogo.domain.entity.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+
 
 @Repository
 public interface MenuRepository extends JpaRepository<Menu, String> {
