+++ domain/src/main/java/com/egomogo/domain/type/CategoryNode.java
@@ -1,4 +1,4 @@
-package com.egomogo.api.service.type;
+package com.egomogo.domain.type;
 
 import lombok.Getter;
 
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
+
 @Getter
 public class CategoryNode {
     private final CategoryType value;
