+++ settings.gradle
@@ -1,2 +1,3 @@
 rootProject.name = 'backend'
-include("api")
+include 'api'
+include 'domain'
